requirements.txt
eth-brownie >= 1.18
